+++ clients/admin-ui/src/features/datastore-connections/constants.ts
@@ -1,4 +1,58 @@
-import { ConnectionType } from "./types";
+/**
+ * Enums
+ */
+export enum AccessLevel {
+  READ = "read",
+  WRITE = "write",
+}
+
+export enum ConnectionTestStatus {
+  SUCCEEDED = "succeeded",
+  FAILED = "failed",
+  SKIPPED = "skipped",
+}
+
+export enum ConnectionType {
+  POSTGRES = "postgres",
+  MONGODB = "mongodb",
+  MYSQL = "mysql",
+  HTTPS = "https",
+  SAAS = "saas",
+  REDSHIFT = "redshift",
+  SNOWFLAKE = "snowflake",
+  MSSQL = "mssql",
+  MARIADB = "mariadb",
+  BIGQUERY = "bigquery",
+  MANUAL = "manual",
+}
+
+export enum DisabledStatus {
+  ACTIVE = "active",
+  DISABLED = "disabled",
+}
+
+export enum SassType {
+  MAILCHIMP = "mailchimp",
+  HUB_SPOT = "hubspot",
+  OUTREACH = "outreach",
+  SEGMENT = "segment",
+  SENTRY = "sentry",
+  STRIPE = "stripe",
+  ZENDESK = "zendesk",
+  CUSTOM = "custom",
+}
+
+export enum SystemType {
+  SAAS = "saas",
+  DATABASE = "database",
+  MANUAL = "manual",
+}
+
+export enum TestingStatus {
+  PASSED = "passed",
+  FAILED = "failed",
+  UNTESTED = "untested",
+}
 
 /**
  * Relative folder path for connector images